Move cmd functions to utils to reuse
