+++ ckanext/qa/commands.py
@@ -71,6 +71,7 @@ class QACommand(p.toolkit.CkanCommand):
             return
 
         cmd = self.args[0]
+        args = self.args[1:] if len(self.args) > 1 else []
         self._load_config()
 
         # Now we can import ckan and create logger, knowing that loggers
@@ -78,7 +79,7 @@ class QACommand(p.toolkit.CkanCommand):
         self.log = logging.getLogger('ckanext.qa')
 
         if cmd == 'update':
-            self.update()
+            utils.update(*args, queue=self.options.queue)
         elif cmd == 'sniff':
             self.sniff()
         elif cmd == 'view':
@@ -94,81 +95,6 @@ class QACommand(p.toolkit.CkanCommand):
             utils.init_db()
         else:
             self.log.error('Command "%s" not recognized' % (cmd,))
-
-    def update(self):
-        from ckan import model
-        from ckanext.qa import lib
-        packages = []
-        resources = []
-        if len(self.args) > 1:
-            for arg in self.args[1:]:
-                # try arg as a group id/name
-                group = model.Group.get(arg)
-                if group and group.is_organization:
-                    # group.packages() is unreliable for an organization -
-                    # member objects are not definitive whereas owner_org, so
-                    # get packages using owner_org
-                    query = model.Session.query(model.Package)\
-                        .filter(
-                            or_(model.Package.state == 'active',
-                                model.Package.state == 'pending'))\
-                        .filter_by(owner_org=group.id)
-                    packages.extend(query.all())
-                    if not self.options.queue:
-                        self.options.queue = 'bulk'
-                    continue
-                elif group:
-                    packages.extend(group.packages())
-                    if not self.options.queue:
-                        self.options.queue = 'bulk'
-                    continue
-                # try arg as a package id/name
-                pkg = model.Package.get(arg)
-                if pkg:
-                    packages.append(pkg)
-                    if not self.options.queue:
-                        self.options.queue = 'priority'
-                    continue
-                # try arg as a resource id
-                res = model.Resource.get(arg)
-                if res:
-                    resources.append(res)
-                    if not self.options.queue:
-                        self.options.queue = 'priority'
-                    continue
-                else:
-                    self.log.error('Could not recognize as a group, package '
-                                   'or resource: %r', arg)
-                    sys.exit(1)
-        else:
-            # all packages
-            pkgs = model.Session.query(model.Package)\
-                        .filter_by(state='active')\
-                        .order_by('name').all()
-            packages.extend(pkgs)
-            if not self.options.queue:
-                self.options.queue = 'bulk'
-
-        if packages:
-            self.log.info('Datasets to QA: %d', len(packages))
-        if resources:
-            self.log.info('Resources to QA: %d', len(resources))
-        if not (packages or resources):
-            self.log.error('No datasets or resources to process')
-            sys.exit(1)
-
-        self.log.info('Queue: %s', self.options.queue)
-        for package in packages:
-            lib.create_qa_update_package_task(package, self.options.queue)
-            self.log.info('Queuing dataset %s (%s resources)',
-                          package.name, len(package.resources))
-
-        for resource in resources:
-            package = resource.resource_group.package
-            self.log.info('Queuing resource %s/%s', package.name, resource.id)
-            lib.create_qa_update_task(resource, self.options.queue)
-
-        self.log.info('Completed queueing')
 
     def sniff(self):
         from ckanext.qa.sniff_format import sniff_file_format
